+++ Kite-SDK/PSPrintSDK/ViewController.m
@@ -60,8 +60,6 @@ static NSString *const kURLScheme = @"kitely";
 - (void)viewDidLoad {
     [super viewDidLoad];
     
-    [OLKitePrintSDK setInstagramEnabledWithClientID:@"1af4c208cbdc4d09bbe251704990638f" secret:@"c8a5b1b1806f4586afad2f277cee1d5c" redirectURI:@"https://kite.ly/instagram-callback"];
-    
     [OLKitePrintSDK setApplePayMerchantID:kApplePayMerchantIDKey];
     [OLKitePrintSDK setApplePayPayToString:kApplePayBusinessName];
     [OLKitePrintSDK setURLScheme:kURLScheme];